Add files via upload
A little script to organice and clean the excel files from the spanish census, from year 2000 to the most actual data.

Files from year 96, 98 and 99 are just removed.  year 97 is not available and this is just an exercise.  

Including them will suposse a increase in the complexity of the script for just 3 files that  are not even consecutives.  


For all the other files, this script will upgrade the .xsl files to .xlsx.

After that it will rename the file, using the year of the data, and finally it will remove the upper lines that hace some unwanted comments.

Finally it will make some changes in the column names.